requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
peft>=0.6.0
bitsandbytes>=0.41.0
accelerate>=0.24.0
huggingface-hub>=0.17.0
python-dotenv>=1.0.0

